driver_state.cpp
testeststets

+++ proj-gl/proj-gl-files/driver_state.cpp
@@ -191,9 +191,9 @@ void rasterize_triangle(driver_state& state, const data_geometry* in[3])
                 
                 state.fragment_shader(fragment_data, out_data, state.uniform_data);
                 
-                state.image_color[i + j * state.image_width] = make_pixel(static_cast<int>(out_data.output_color[0] * 255),
-                                                                          static_cast<int>(out_data.output_color[1] * 255),
-                                                                          static_cast<int>(out_data.output_color[2] * 255));
+                state.image_color[i + j * state.image_width] = make_pixel((out_data.output_color[0] * 255),
+                                                                          (out_data.output_color[1] * 255),
+                                                                          (out_data.output_color[2] * 255));
             }
         }
     }
